Index

SolMod.errorAnalysis

SolMod.idealActivityCoefficient

SolMod.ternaryPhase

Base Functions

SolMod.idealActivityCoefficient — Function

idealActivityCoefficient(; x::Vector = [0.3, 0.2, 0.5])

The function will always output an activity coefficient of 1

Optional:

x = Vector that includes the composition of the soluion. Automatically set to [0.3, 0.2, 0.5]

Missing docstring.

Missing docstring for SolMod.ImportSolubility. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SolMod.ExportSolubility. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SolMod.ImportTernaryPhase. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SolMod.ExportTernaryPhase. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SolMod.ImportExperimentalSolubility. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SolMod.ExportExperimentalSolubility. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SolMod.ImportExperimentalTernaryPhase. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SolMod.ExportExperimentalTernaryPhase. Check Documenter's build log for details.

SolMod.ternaryPhase — Function

ternaryPhase(params::Dict, γi, γj, Tx; x_step::Float64 = 0.001, 
    x_start::Float64 = 0.001, x_end::Float64 = 0.999, round1::Int64 = 3, round2::Int64 = 3, 
    round3::Int64 = 3, pp::Bool = true, e::Bool = true)

The function will find the activity coefficient of the R enantiomer and the S enantiomer and checks if the Schrödenberg Van Laar equation and the Prigogine and Defay equation is correct. This is done for every molar composition possible. When the solubility equation is true, the molar composition is appended to an array. The array is then processed to remove solubility points beyond the eutectic points

Inputs:

params = Dictionary that includes the solvents used as keys and the respective

interaction parameters. The solute key contains the calorimetric data of the target molecule as an array

solvent = String of the solvent used for the modeling

γix = Activtiy coefficient of the i component at temperature Tx in the x compositon

γjx = Activtiy coefficient of the j component at temperature Tx in the x compositon

Tx = Value of the type Int64/Float64 of the temperature used to determine the

activity coefficient

Optional:

x_step = Step size of the molar composition increments. Automatically set to 0.001

x_start = Start of the iteration of the molar composition. Automatically set to 0.001

x_end = End of the iteration of the molar composition. Automatically set to 0.999

round# = This affects the sensitivity of the solubility equation. This rounds to the

specified number of digits after the decimal place. Automatically set to 4 for the solubility of the enantiomers and 3 for the solubility of the racemate

pp = Boolean that allows the post-processing of the output array will be done to

remove any points beyond the eutectic points

Missing docstring.

Missing docstring for SolMod.solubilityPoint. Check Documenter's build log for details.

SolMod.errorAnalysis — Function